+++ lib/ec2launcher/terminator.rb
@@ -7,6 +7,7 @@ require 'log4r'
 
 require 'ec2launcher/aws_initializer'
 require 'ec2launcher/backoff_runner'
+require 'ec2launcher/route53'
 
 module EC2Launcher
   class Terminator
@@ -22,10 +23,37 @@ module EC2Launcher
 
     def terminate(server_name, options)
       ##############################
+      # Load configuration data
+      ##############################
+      config_wrapper = ConfigWrapper.new(options.directory)
+
+      @config = config_wrapper.config
+      @environments = config_wrapper.environments
+
+      ##############################
+      # ENVIRONMENT
+      ##############################
+      unless @environments.has_key? options.environ
+        @log.fatal "Environment not found: #{options.environ}"
+        exit 2
+      end
+      @environment = @environments[options.environ]
+      
+      ##############################
       # Initialize AWS and create EC2 connection
       ##############################
       initialize_aws(options.access_key, options.secret)
       @ec2 = AWS::EC2.new
+      
+      ##############################
+      # Create Route53 connection
+      ##############################
+      aws_route53 = AWS::Route53.new if @environment.route53_zone_id
+      @route53 = EC2Launcher::Route53.new(aws_route53, @environment.route53_zone_id)
+
+      ##############################
+      # Find instance
+      ##############################
       instance = nil
       AWS.memoize do
         instances = @ec2.instances.filter("tag:Name", server_name)
@@ -40,6 +68,12 @@ module EC2Launcher
       if instance
         @log.info("Terminating instance: #{server_name} [#{instance.instance_id}]")
         instance.terminate
+
+        if @route53
+          @log.info("Deleting A record from Route53: #{server_name} => #{instance.private_ip_address}")
+          @route53.delete_record_by_name(server_name, 'A')
+        end
+
         @log.info("Deleting node/client from Chef: #{server_name}")
         node_result = `echo "Y" |knife node delete #{server_name}`
         client_result = `echo "Y" |knife client delete #{server_name}`